requirements.txt
pandas
matplotlib
numpy
skaero.atmosphere
scipy
scipy.stats
pygps
gpxpy
pysat
pysal
